+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to NebulaOS will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [0.0.3] - In Progress
+### Added
+- Added basic audio support via AC97
+- Added PCI support
+- Added various UI improvements
 
 ## [0.0.3-dev3] - 2026-05-01
 ### Changed